[문자열 계산기] 반환형 double 로 변경

+++ src/main/java/calculator/MainClass.java
@@ -9,7 +9,7 @@ public class MainClass {
         StringCalculator stringCalculator = new StringCalculator();
         String input = scanner.nextLine();
 
-        long result = stringCalculator.calculte(input);
+        double result = stringCalculator.calculte(input);
         System.out.println(result);
     }
 }
